+++ app/scripts/controllers/network/createJsonRpcClient.js
@@ -1,10 +1,7 @@
 const mergeMiddleware = require('json-rpc-engine/src/mergeMiddleware')
-const createFetchMiddleware = require('eth-json-rpc-middleware/fetch')
-const createBlockRefMiddleware = require('eth-json-rpc-middleware/block-ref')
-const createBlockCacheMiddleware = require('eth-json-rpc-middleware/block-cache')
-const createInflightMiddleware = require('eth-json-rpc-middleware/inflight-cache')
-const createBlockTrackerInspectorMiddleware = require('eth-json-rpc-middleware/block-tracker-inspector')
-const providerFromMiddleware = require('eth-json-rpc-middleware/providerFromMiddleware')
+const createFetchMiddleware = require('bch-json-rpc-middleware/fetch')
+const createInflightMiddleware = require('bch-json-rpc-middleware/inflight-cache')
+const providerFromMiddleware = require('bch-json-rpc-middleware/providerFromMiddleware')
 const BlockTracker = require('eth-block-tracker')
 
 module.exports = createJsonRpcClient