+++ app/scripts/metamask-controller.js
@@ -51,6 +51,8 @@ const EthQuery = require('eth-query')
 const ethUtil = require('ethereumjs-util')
 // const sigUtil = require('eth-sig-util')
 const axios = require('axios')
+const CreateSlpSdk = require('slp-sdk')
+const SlpSdk = new CreateSlpSdk()
 
 module.exports = class MetamaskController extends EventEmitter {
   /**
@@ -268,6 +270,16 @@ module.exports = class MetamaskController extends EventEmitter {
           return []
         }
       },
+      getSlpAccounts: async () => {
+        const isUnlocked = this.keyringController.memStore.getState().isUnlocked
+        const selectedAddress = this.preferencesController.getSelectedSlpAddress()
+        // only show address if account is unlocked
+        if (isUnlocked && selectedAddress) {
+          return [selectedAddress]
+        } else {
+          return []
+        }
+      },
       // tx signing
       processTransaction: this.newUnapprovedTransaction.bind(this),
       // msg signing
@@ -299,6 +311,9 @@ module.exports = class MetamaskController extends EventEmitter {
       const result = {
         selectedAddress: memState.isUnlocked
           ? memState.selectedAddress
+          : undefined,
+        selectedSlpAddress: memState.isUnlocked
+          ? SlpSdk.Address.toSLPAddress(memState.selectedSlpAddress)
           : undefined,
         networkVersion: memState.network,
       }
@@ -369,7 +384,10 @@ module.exports = class MetamaskController extends EventEmitter {
       clearSeedWordCache: this.clearSeedWordCache.bind(this),
       resetAccount: nodeify(this.resetAccount, this),
       removeAccount: nodeify(this.removeAccount, this),
-      importAccountWithStrategy: nodeify(this.importAccountWithStrategy, this),
+      importAccountWithStrategy: nodeify(
+        this.importAccountWithStrategy,
+        this
+      ),
 
       // hardware wallets
       connectHardware: nodeify(this.connectHardware, this),
@@ -383,6 +401,8 @@ module.exports = class MetamaskController extends EventEmitter {
       // vault management
       submitPassword: nodeify(this.submitPassword, this),
 
+      checkVaultEncrypted: nodeify(this.checkVaultEncrypted, this),
+
       // network management
       setProviderType: nodeify(
         networkController.setProviderType,
@@ -395,7 +415,10 @@ module.exports = class MetamaskController extends EventEmitter {
         preferencesController.setSelectedAddress,
         preferencesController
       ),
-      addToken: nodeify(preferencesController.addToken, preferencesController),
+      addToken: nodeify(
+        preferencesController.addToken,
+        preferencesController
+      ),
       removeToken: nodeify(
         preferencesController.removeToken,
         preferencesController
@@ -425,8 +448,14 @@ module.exports = class MetamaskController extends EventEmitter {
 
       // KeyringController
       setLocked: nodeify(keyringController.setLocked, keyringController),
-      createNewVaultAndKeychain: nodeify(this.createNewVaultAndKeychain, this),
-      createNewVaultAndRestore: nodeify(this.createNewVaultAndRestore, this),
+      createNewVaultAndKeychain: nodeify(
+        this.createNewVaultAndKeychain,
+        this
+      ),
+      createNewVaultAndRestore: nodeify(
+        this.createNewVaultAndRestore,
+        this
+      ),
       addNewKeyring: nodeify(
         keyringController.addNewKeyring,
         keyringController
@@ -437,14 +466,23 @@ module.exports = class MetamaskController extends EventEmitter {
       ),
 
       // txController
-      cancelTransaction: nodeify(txController.cancelTransaction, txController),
-      updateTransaction: nodeify(txController.updateTransaction, txController),
+      cancelTransaction: nodeify(
+        txController.cancelTransaction,
+        txController
+      ),
+      updateTransaction: nodeify(
+        txController.updateTransaction,
+        txController
+      ),
       updateAndApproveTransaction: nodeify(
         txController.updateAndApproveTransaction,
         txController
       ),
       retryTransaction: nodeify(this.retryTransaction, this),
-      getFilteredTxList: nodeify(txController.getFilteredTxList, txController),
+      getFilteredTxList: nodeify(
+        txController.getFilteredTxList,
+        txController
+      ),
       isNonceTaken: nodeify(txController.isNonceTaken, txController),
       estimateGas: nodeify(this.estimateGas, this),
 
@@ -461,8 +499,12 @@ module.exports = class MetamaskController extends EventEmitter {
       cancelTypedMessage: this.cancelTypedMessage.bind(this),
 
       // notices
-      checkNotices: noticeController.updateNoticesList.bind(noticeController),
-      markNoticeRead: noticeController.markNoticeRead.bind(noticeController),
+      checkNotices: noticeController.updateNoticesList.bind(
+        noticeController
+      ),
+      markNoticeRead: noticeController.markNoticeRead.bind(
+        noticeController
+      ),
     }
   }
 
@@ -488,13 +530,15 @@ module.exports = class MetamaskController extends EventEmitter {
     const releaseLock = await this.createVaultMutex.acquire()
     try {
       let vault
-      const accounts = await this.keyringController.getAccounts()
+      let accounts = await this.keyringController.getAccounts()
+      let slpAccounts = await this.keyringController.getSlpAccounts()
       if (accounts.length > 0) {
         vault = await this.keyringController.fullUpdate()
       } else {
         vault = await this.keyringController.createNewVaultAndKeychain(password)
-        const accounts = await this.keyringController.getAccounts()
-        this.preferencesController.setAddresses(accounts)
+        accounts = await this.keyringController.getAccounts()
+        slpAccounts = await this.keyringController.getSlpAccounts()
+        this.preferencesController.setAddresses(accounts, slpAccounts)
         this.selectFirstIdentity()
       }
       releaseLock()
@@ -516,7 +560,7 @@ module.exports = class MetamaskController extends EventEmitter {
       const keyringController = this.keyringController
 
       // clear known identities
-      this.preferencesController.setAddresses([])
+      this.preferencesController.setAddresses([], [])
       // create new vault
       const vault = await keyringController.createNewVaultAndRestore(
         password,
@@ -524,6 +568,7 @@ module.exports = class MetamaskController extends EventEmitter {
       )
 
       const accounts = await keyringController.getAccounts()
+      const slpAccounts = await keyringController.getSlpAccounts()
 
       const primaryKeyring = keyringController.getKeyringsByType(
         'HD Key Tree'
@@ -533,7 +578,7 @@ module.exports = class MetamaskController extends EventEmitter {
       }
 
       // set new identities
-      this.preferencesController.setAddresses(accounts)
+      this.preferencesController.setAddresses(accounts, slpAccounts)
       this.selectFirstIdentity()
       releaseLock()
       return vault
@@ -577,6 +622,7 @@ module.exports = class MetamaskController extends EventEmitter {
   async submitPassword (password) {
     await this.keyringController.submitPassword(password)
     const accounts = await this.keyringController.getAccounts()
+    const slpAccounts = await this.keyringController.getSlpAccounts()
 
     // verify keyrings
     const nonSimpleKeyrings = this.keyringController.keyrings.filter(
@@ -586,7 +632,12 @@ module.exports = class MetamaskController extends EventEmitter {
       await this.diagnostics.reportMultipleKeyrings(nonSimpleKeyrings)
     }
 
-    await this.preferencesController.syncAddresses(accounts)
+    await this.preferencesController.setAddresses(accounts, slpAccounts)
+    return this.keyringController.fullUpdate()
+  }
+
+  async checkVaultEncrypted (password) {
+    await this.keyringController.submitPassword(password)
     return this.keyringController.fullUpdate()
   }
 
@@ -742,10 +793,11 @@ module.exports = class MetamaskController extends EventEmitter {
     const oldAccounts = await keyringController.getAccounts()
     const keyState = await keyringController.addNewAccount(primaryKeyring)
     const newAccounts = await keyringController.getAccounts()
+    const slpAccounts = await keyringController.getSlpAccounts()
 
     await this.verifySeedPhrase()
 
-    this.preferencesController.setAddresses(newAccounts)
+    this.preferencesController.setAddresses(newAccounts, slpAccounts)
     newAccounts.forEach(address => {
       if (!oldAccounts.includes(address)) {
         this.preferencesController.setSelectedAddress(address)
@@ -831,8 +883,6 @@ module.exports = class MetamaskController extends EventEmitter {
   async resetAccount () {
     const selectedAddress = this.preferencesController.getSelectedAddress()
     this.txController.wipeTransactions(selectedAddress)
-    this.networkController.resetConnection()
-
     return selectedAddress
   }
 
@@ -871,7 +921,8 @@ module.exports = class MetamaskController extends EventEmitter {
     const accounts = await keyring.getAccounts()
     // update accounts in preferences controller
     const allAccounts = await this.keyringController.getAccounts()
-    this.preferencesController.setAddresses(allAccounts)
+    const allSlpAccounts = await this.keyringController.getSlpAccounts()
+    this.preferencesController.setAddresses(allAccounts, allSlpAccounts)
     // set new account as selected
     await this.preferencesController.setSelectedAddress(accounts[0])
   }
@@ -1352,13 +1403,14 @@ module.exports = class MetamaskController extends EventEmitter {
       (acc, { accounts }) => acc.concat(accounts),
       []
     )
+    const slpAddresses = await this.keyringController.getSlpAccounts()
 
     if (!addresses.length) {
       return
     }
 
     // Ensure preferences + identities controller know about all addresses
-    this.preferencesController.addAddresses(addresses)
+    this.preferencesController.addAddresses(addresses, slpAddresses)
     this.accountTracker.syncWithAddresses(addresses)
 
     const wasLocked = !isUnlocked