+++ ui/app/components/app-header/app-header.component.js
@@ -25,12 +25,17 @@ export default class AppHeader extends PureComponent {
     showNetworkDropdown: PropTypes.func,
     hideNetworkDropdown: PropTypes.func,
     toggleAccountMenu: PropTypes.func,
+    checkUnencrypted: PropTypes.func,
     selectedAddress: PropTypes.string,
     isUnlocked: PropTypes.bool,
   }
 
   static contextTypes = {
     t: PropTypes.func,
+  }
+  componentDidMount () {
+    const { checkUnencrypted } = this.props
+    checkUnencrypted()
   }
 
   handleNetworkIndicatorClick (event) {
@@ -103,11 +108,11 @@ export default class AppHeader extends PureComponent {
   }
 
   render () {
-    const { network, provider, history, isUnlocked } = this.props
+    const { history, isUnlocked } = this.props
 
-    if (this.hideAppHeader()) {
-      return null
-    }
+    // if (this.hideAppHeader()) {
+    //   return null
+    // }
 
     return (
       <div